backend/doc_processor/requirements.txt
Flask
Flask-Cors
PyMuPDF
faiss-cpu
sentence-transformers
numpy
tf-keras
protobuf
